Moving CompliantReviewsGrid to separate file
